backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic-ai-slim[google]==0.5.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
pandas>=2.0.0
nest_asyncio>=1.5.0
tqdm>=4.66.0
google-genai>=1.24.0
